COM-117: configuration persists after closing the application
Added server file system interface, which can handle post requests to save website appearance configuration info.
I've throw together a quick mockup of the form in the admin-configuration page.
Note that I am still in the process of finalizing this design, as it will likely change to accomodate loading the config.